allow configuration via preloaded captures
Using captures as configuration to drive test execution.

The captures have to be set up before the tests are loaded.

This is problematic since captures were stored against the fixture
object; so we change that to use the class of the fixture object
instead, which means a simple __PACKAGE__->capture(...) is usually all
that is required.

That changes also means that captures can communicate across test
documents, so care is needed where collisions are possible.

It might be better to use a different syntax, e.g. \c<key> for
class-level captures and keep using \k<key> for per-fixture captures.