support 'ncmd' commands which are not tests
Almost allows non-test commands to be specified, e.g. for preparation
or cleanup purposes. Captures and parsers are still executed.

It's only "almost" because Test::Command::Simple insists on testing
the startup of the command (return code from open3), so there is still
one test associated with an 'ncmd' test doc.